+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fix `incident_catalog_type` to be able to better handle undefined or empty category lists
 - Expose `incident_catalog_entries` in the documentation
 - No longer escape HTML characters in engine JSON strings
+- `incident_alert_attribute` now has an optional `required` property. Set this
+  to true for attributes which you expect to be present on all alerts.
 
 ## v5.11.0